figured out the bullet issue I had -finished
i think i finished the game, still has some minor bugs which can be
fixed, bullets work and attack montsers - as well as decreasing their
health.

The monsters are dying and a total of 4 waves are occuring making the
game fun and short.

Other elements can be added later on such as the bugs attacking the
towers making it more challenging for the player